requirements.txt
openai==1.59.7
urllib3==2.2.3
langchain==0.3.17
langchain-anthropic==0.3.5
langchain-community==0.3.16
langchain-core==0.3.33
langchain-openai==0.3.0
langchain-redis==0.1.1
langchain-text-splitters==0.3.5
langgraph==0.2.68
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
python-dotenv==1.0.1
pydantic==2.10.1
requests==2.32.3
langgraph==0.2.68
openpyxl==3.1.5
pandas==2.2.3
bs4==0.0.2
chardet==5.2.0
charset-normalizer==3.4.1